+++ src/main/java/be/ugent/knows/idlabFunctions/state/MapDBContainer.java
@@ -4,7 +4,10 @@ import org.mapdb.DB;
 import org.mapdb.DBMaker;
 import org.mapdb.HTreeMap;
 import org.mapdb.Serializer;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import java.io.File;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
@@ -16,17 +19,37 @@ import java.util.concurrent.TimeUnit;
  * @author Gerald Haesendonck
  */
 public class MapDBContainer {
+    private final static Logger logger = LoggerFactory.getLogger(MapDBContainer.class);
     private final DB mapDB;
     private final HTreeMap<String, String> map;
 
     private final ScheduledExecutorService committer;
 
+    /**
+     * Creates a MapDBContainer instance. If {@code dbFilePath} is given, it is backed by a memory mapped mapdb.
+     * If {@code dbFilePath} is {@code null} or cannot be created, a temporary file is created and deleted when closed.
+     * @param dbFilePath    The path to the mapdb file. If it exists, the file is loaded. If it doesn't, a new map is created.
+     */
     public MapDBContainer(final String dbFilePath) {
-        mapDB = DBMaker
-                .fileDB(dbFilePath)
-                .fileMmapEnableIfSupported()
+
+        DBMaker.Maker dbMaker;
+        if (dbFilePath != null) {
+            File dbPArentFile = new File(dbFilePath).getParentFile();
+            if (!dbPArentFile.exists() && !dbPArentFile.mkdirs()) {
+                logger.warn(dbPArentFile + " doesn not exist and could not be created, creating a temporary file state.");
+                dbMaker = DBMaker.tempFileDB();
+            } else {
+                dbMaker = DBMaker.fileDB(dbFilePath);
+            }
+        } else {
+            logger.debug("dbFilePath is null, creating a temporary state.");
+            dbMaker = DBMaker.tempFileDB();
+        }
+
+        mapDB = dbMaker.fileMmapEnableIfSupported()
                 .closeOnJvmShutdown()
                 .make();
+
         map = mapDB.hashMap("map", Serializer.STRING, Serializer.STRING).createOrOpen();
 
         // make an executor service that does a commit on the mapDB every 10 seconds